+++ src/dropdownToggle/dropdownToggle.js
@@ -28,10 +28,14 @@ angular.module('mm.foundation.dropdownToggle', [ 'mm.foundation.position', 'mm.f
     },
     controller: 'DropdownToggleController',
     link: function(scope, element, attrs, controller) {
+      var parent = element.parent();
       var dropdown = angular.element($document[0].querySelector(scope.dropdownToggle));
 
-      scope.$watch('$location.path', function() { closeMenu(); });
-      element.bind('click', function (event) {
+      var parentHasDropdown = function() {
+        return parent.hasClass('has-dropdown');
+      };
+
+      var onClick = function (event) {
         dropdown = angular.element($document[0].querySelector(scope.dropdownToggle));
         var elementWasOpen = (element === openElement);
 
@@ -43,15 +47,12 @@ angular.module('mm.foundation.dropdownToggle', [ 'mm.foundation.position', 'mm.f
         }
 
         if (!elementWasOpen && !element.hasClass('disabled') && !element.prop('disabled')) {
-          dropdown.css('display', 'block');
-
+          dropdown.css('display', 'block'); // We display the element so that offsetParent is populated
           var offset = $position.offset(element);
           var parentOffset = $position.offset(angular.element(dropdown[0].offsetParent));
-
           var dropdownWidth = dropdown.prop('offsetWidth');
-
           var css = {
-              top: offset.top - parentOffset.top + offset.height + 'px'
+            top: offset.top - parentOffset.top + offset.height + 'px'
           };
 
           if (controller.small()) {
@@ -74,20 +75,35 @@ angular.module('mm.foundation.dropdownToggle', [ 'mm.foundation.position', 'mm.f
 
           dropdown.css(css);
 
+          if (parentHasDropdown()) {
+            parent.addClass('hover');
+          }
+
           openElement = element;
+
           closeMenu = function (event) {
-            $document.unbind('click', closeMenu);
+            $document.off('click', closeMenu);
             dropdown.css('display', 'none');
             closeMenu = angular.noop;
             openElement = null;
+            if (parent.hasClass('hover')) {
+              parent.removeClass('hover');
+            }
           };
-          $document.bind('click', closeMenu);
+          $document.on('click', closeMenu);
         }
-      });
+      };
 
       if (dropdown) {
         dropdown.css('display', 'none');
       }
+
+      scope.$watch('$location.path', function() { closeMenu(); });
+
+      element.on('click', onClick);
+      element.on('$destroy', function() {
+        element.off('click', onClick);
+      });
     }
   };
 }]);